+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **fixed:** In `Reflect::to_value` for enums, only generate a catch all branch
   if the enum has a variant with `#[reflect(skip)]` ([#105])
+- **added:** Add a `has_default_value` method to types in `type_info` ([#104])
 
 [#105]: https://github.com/EmbarkStudios/mirror-mirror/pull/105
+[#104]: https://github.com/EmbarkStudios/mirror-mirror/pull/104
 
 # 0.1.8 (23. February, 2023)